requirements.txt
streamlit
pandas
Pillow
folium
streamlit-folium